added help button, made landing page go to login page, added a minimum size to the toolbar to help with different screen sizes, added emojis to toolbar. Also made toolbar a component and moved components to src
