+++ src/app/auth/page.js
@@ -15,7 +15,7 @@ import {
 } from 'firebase/auth';
 
 import { useRouter } from 'next/navigation';
-import Toolbar from '../Toolbar';
+import Toolbar from '../components/Toolbar';
 
 export default function Login() {
   const [error, setError] = useState('');
