+++ src/app/main-menu/page.js
@@ -2,7 +2,7 @@
 
 import React from 'react';
 import Link from 'next/link';
-import Toolbar from '../Toolbar';
+import Toolbar from '../components/Toolbar';
 
 export default function MainMenu() {
   const features = [
