+++ src/app/quizzes/page.js
@@ -1,7 +1,7 @@
 'use client';
 
 import React, { useState, useEffect } from 'react';
-import Toolbar from '../Toolbar';
+import Toolbar from '../components/Toolbar';
 import Link from 'next/link';
 import { useAuth } from '../context/AuthContext';
 import { firebaseDB } from '../../../firebase_config';
@@ -124,6 +124,12 @@ export default function Quizzes() {
           </div>
         </div>
       )}
+        <button
+          onClick={() => setShowPopup(true)}
+          className="bg-black text-white p-4 rounded-full shadow-lg hover:bg-green-600 hover:text-black fixed top-4 right-4 flex items-center justify-center w-16 h-16"
+        >
+          ❔
+        </button>
 
       <main className="flex flex-col gap-8 row-start-2 items-center justify-center sm:items-center max-w-lg w-full">
         {quiz ? (
