+++ src/app/text-to-speech/page.js
@@ -1,8 +1,8 @@
 'use client';
 
 import React, { useState } from 'react';
-import Toolbar from '../Toolbar';
-import TextToSpeechButton from '../../../components/TextToSpeechButton';
+import Toolbar from '../components/Toolbar';
+import TextToSpeechButton from '../components/TextToSpeechButton';
 
 export default function TextToSpeech() {
   const [showPopup, setShowPopup] = useState(true);
@@ -28,6 +28,12 @@ export default function TextToSpeech() {
           </div>
         </div>
       )}
+        <button
+          onClick={() => setShowPopup(true)}
+          className="bg-black text-white p-4 rounded-full shadow-lg hover:bg-green-600 hover:text-black fixed top-4 right-4 flex items-center justify-center w-16 h-16"
+        >
+          ❔
+        </button>
 
       <div
         data-testid="Text to Speech"
